requirements.txt

aiofiles==23.2.1
python-dotenv==0.20.0